Parse incoming mpegts data to improve delay queue behavior
With srtsrc linked directly to the delay queue, I observed the
current-level-time metric go down while the current-level-bytes metric
went up. I suspect that the timestamp information from srtsrc is not
sufficient to be used as a basis for time-based buffering. Running the
data through tsparse first appears to fix this problem.