Use funnel instead of input selector to maintain continuity
This is a fix to a nasty bug I've been working on the past few days:
when entering censored state, the video queue empties out and then
there's a stream discontinuity. The un-delayed SRT feed stalled at the
beginning of the censorship, and the delayed video feed stalled at the
end of the censorship period. I think what might be happening is that
there's a moment where the pipeline is not connected when we're swapping
the input and output selectors. I tried reversing the order of updating
the output and input selector pads, but that didn't work. Using a funnel
connected to both always does.

There actually used to be a funnel here, but I removed it in
7bde5e1278de6ee3ceb22df85954081a22a67e8d due to it causing flickers. I
haven't been able to reproduce this -- yet?